+++ tests/testthat/test-SkewMacdonaldPolynomial.R
@@ -3,7 +3,7 @@ test_that("Skew Macdonald at q=0 is skew Hall-Littlewood - P", {
   lambda <- c(3, 2)
   mu <- c(1)
   hlPoly <- SkewHallLittlewoodPol(n, lambda, mu, "P")
-  macPoly <- SkewMacdonaldPolynomial(n, lambda, mu, "P")
+  macPoly <- SkewMacdonaldPol(n, lambda, mu, "P")
   coeffs <- lapply(macPoly@coeffs, function(rOQ) {
     swapVariables(
       substituteRatioOfQsprays(rOQ, values = c(0, NA)),
@@ -23,7 +23,7 @@ test_that("Skew Macdonald at q=0 is skew Hall-Littlewood - Q", {
   lambda <- c(3, 2)
   mu <- c(1, 1)
   hlPoly <- SkewHallLittlewoodPol(n, lambda, mu, "Q")
-  macPoly <- SkewMacdonaldPolynomial(n, lambda, mu, "Q")
+  macPoly <- SkewMacdonaldPol(n, lambda, mu, "Q")
   coeffs <- lapply(macPoly@coeffs, function(rOQ) {
     swapVariables(
       substituteRatioOfQsprays(rOQ, values = c(0, NA)),