+++ NEWS.md
@@ -35,6 +35,10 @@ polynomial. A skew Hall-Littlewood polynomial is a multivariate symmetric
 polynomial associated to a skew integer partition and depending on a parameter. 
 When the value of this parameter is 0, then this is the skew Schur polynomial.
 
+- New function `MacdonaldPol`, to get a Macdonald polynomial.
+
+- New function `SkewMacdonaldPol`, to get a skew Macdonald polynomial.
+
 - New function `flaggedSchurPol`, to get a flagged Schur polynomial.
 
 - New function `flaggedSkewSchurPol`, to get a flagged skew Schur polynomial.